+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.12.3] 2025-07-15
+
+### Fixed
+
+- GUI_tools: 
+    - `ConfigManager` and `GUIFactory` will correctly exchange configuration details.
+    - `GUIFactory.generate_continuous_layout()` will not cast range values to integers, it will use the same type as received.
+
 ## [3.12.2] 2025-07-15
 
 ### Fixed